refactor: change type.ts to types.ts
