+++ apps/web/src/modules/job/detail/components/JobCard/components/JobCardHeader/types.ts
