+++ apps/web/src/modules/job/list/components/JobCardContainer/components/JobCard/types.ts
