+++ apps/web/src/modules/job/list/components/NotiCardContainer/components/NotiCard/index.tsx
@@ -4,7 +4,7 @@ import Footer from './components/Footer'
 import Header from './components/Header'
 import OfferDetail from './components/OfferDetail'
 import { CardContainer } from './styled'
-import { CardProps } from './type'
+import { CardProps } from './types'
 
 export default function NotiCard(prop: CardProps) {
   const { title, companyName, offer, castingImage } = prop
