+++ apps/web/src/modules/job/list/components/NotiCardContainer/types.ts
